Cambios finales V2 a la funcion resta y divicion

+++ resta.c
@@ -1,8 +1,9 @@
 #include <stdio.h>
 
-float resta(int a, int b)
+int resta(int a, int b)
 {
   int c;
   c = a - b;
   return c;
+
 }
